use RUSTC environment var in `parse_and_replace::compile` if present
Cargo respects this environment variable (to specify the path to what
rustc binary to use), but the Rustfix test suite did not. However, this
capability is useful when developing new compiler diagnostics that one
wants Rustfix to be able to handle (this being inspired by the endeavor
that is rust-lang/rust#53013).